serverr/requirements.txt
grpcio
grpcio-tools
alembic
PyMySQL
SQLAlchemy
mysql-connector-python